requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
DateTime==5.3
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
fonttools==4.44.3
gunicorn==21.2.0
holidays==0.36
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
Naked==0.1.32
nelson-siegel-svensson==0.5.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
status==0.2.5
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
zope.interface==6.1
